Move encodeUTF8
