+++ packages/skygear-core/src/index.ts
@@ -7,6 +7,7 @@ export * from "./storage";
 export * from "./url";
 export * from "./imageprocessing";
 export * from "./validation";
+export * from "./utf8";
 /**
  * @public
  */
